Combined NoteForm with NoteCrud
NoteForm is now a part of NoteCrud (well technically still 2 components, but in the same file). This makes the logic more readable, slightly reduces the lines of code, and reduces the number of exports. This should be done for other upserts as well
