+++ packages/ui/src/forms/types.ts
@@ -2,8 +2,9 @@ import { InputType, NoteVersion, Reminder, ReminderCreateInput } from "@local/sh
 import { CodeInputProps as CP, DropzoneProps as DP, IntegerInputProps as QP, LanguageInputProps as LP, SelectorProps as SP, TagSelectorProps as TP } from "components/inputs/types";
 import { FormikProps } from "formik";
 import { MakeLazyRequest } from "hooks/useLazyFetch";
-import { ReactNode } from "react";
+import { Dispatch, ReactNode, SetStateAction } from "react";
 import { Forms } from "utils/consts";
+import { ListObject } from "utils/display/listTools";
 import { ApiVersionShape } from "utils/shape/models/apiVersion";
 import { BookmarkListShape } from "utils/shape/models/bookmarkList";
 import { BotShape } from "utils/shape/models/bot";
@@ -30,6 +31,7 @@ import { ScheduleShape } from "utils/shape/models/schedule";
 import { SmartContractVersionShape } from "utils/shape/models/smartContractVersion";
 import { StandardVersionShape } from "utils/shape/models/standardVersion";
 import { TagShape } from "utils/shape/models/tag";
+import { CrudProps } from "views/objects/types";
 import { ViewDisplayType } from "views/types";
 
 //==============================================================
@@ -57,6 +59,13 @@ export interface BaseObjectFormProps<T> extends FormikProps<T> {
     onCancel: () => unknown;
     onClose: () => unknown;
     ref: React.RefObject<any>;
+}
+
+export interface ImprovedFormProps<Model extends ListObject, ModelShape> extends Omit<CrudProps<Model>, "isLoading">, FormikProps<ModelShape> {
+    disabled: boolean;
+    existing: ModelShape;
+    handleUpdate: Dispatch<SetStateAction<ModelShape>>;
+    isReadLoading: boolean;
 }
 
 export interface BaseGeneratedFormProps {
@@ -101,12 +110,7 @@ export interface NodeRoutineListFormProps extends BaseObjectFormProps<NodeWithRo
 }
 export type FocusModeFormProps = BaseObjectFormProps<FocusModeShape>
 export type MeetingFormProps = BaseObjectFormProps<MeetingShape>
-export interface NoteFormProps extends BaseObjectFormProps<NoteVersionShape> {
-    disabled: boolean;
-    handleClose: (_?: unknown, reason?: "backdropClick" | "escapeKeyDown" | undefined) => void;
-    handleDeleted: (data: NoteVersion) => void;
-    versions: string[];
-}
+export type NoteFormProps = ImprovedFormProps<NoteVersion, NoteVersionShape>
 export type OrganizationFormProps = BaseObjectFormProps<OrganizationShape>
 export interface ProjectFormProps extends BaseObjectFormProps<ProjectVersionShape> {
     versions: string[];
